// src/core/http/generated/models/admin-payment-gateways/paymentGateway.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Payment Gateways
 * Filtered OpenAPI specification for Admin Payment Gateways.
 * OpenAPI spec version: 0.0.1
 */

export interface PaymentGateway {
  id: number;
  name: string;
  /** @nullable */
  provider: string | null;
  /** @nullable */
  status: string | null;
  is_default: boolean;
  is_sandbox: boolean;
  /** @nullable */
  webhook_url: string | null;
  checkout_config: unknown[];
  metadata: unknown[];
  has_public_key: boolean;
  has_access_token: boolean;
  has_client_id: boolean;
  has_client_secret: boolean;
  has_webhook_secret: boolean;
  /** @nullable */
  public_key_masked: string | null;
  /** @nullable */
  last_tested_at: string | null;
  /** @nullable */
  last_test_status: string | null;
  /** @nullable */
  last_test_message: string | null;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}